move the test util to another package
