+++ core/client/fs/src/test/java/alluxio/client/block/options/LockBlockOptionsTest.java
@@ -14,7 +14,7 @@ package alluxio.client.block.options;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import alluxio.CommonTestUtils;
+import alluxio.test.utils.CommonTestUtils;
 import alluxio.thrift.LockBlockTOptions;
 
 import org.junit.Test;
