+++ core/client/fs/src/test/java/alluxio/client/file/options/CreateDirectoryOptionsTest.java
@@ -14,12 +14,12 @@ package alluxio.client.file.options;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 
-import alluxio.CommonTestUtils;
 import alluxio.Configuration;
 import alluxio.Constants;
 import alluxio.PropertyKey;
 import alluxio.client.WriteType;
 import alluxio.security.authorization.Mode;
+import alluxio.test.utils.CommonTestUtils;
 import alluxio.thrift.CreateDirectoryTOptions;
 import alluxio.wire.TtlAction;
