+++ core/client/fs/src/test/java/alluxio/client/file/options/MountOptionsTest.java
@@ -14,7 +14,7 @@ package alluxio.client.file.options;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import alluxio.CommonTestUtils;
+import alluxio.test.utils.CommonTestUtils;
 import alluxio.thrift.MountTOptions;
 
 import org.junit.Test;
