+++ core/client/fs/src/test/java/alluxio/client/lineage/options/DeleteLineageOptionsTest.java
@@ -11,7 +11,7 @@
 
 package alluxio.client.lineage.options;
 
-import alluxio.CommonTestUtils;
+import alluxio.test.utils.CommonTestUtils;
 
 import org.junit.Assert;
 import org.junit.Test;
