+++ core/common/src/test/java/alluxio/test/utils/CommonTestUtilsTest.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio;
+package alluxio.test.utils;
 
 import alluxio.underfs.UnderFileSystem.SpaceType;
