+++ core/common/src/test/java/alluxio/util/WaitForOptionsTest.java
@@ -11,7 +11,7 @@
 
 package alluxio.util;
 
-import alluxio.CommonTestUtils;
+import alluxio.test.utils.CommonTestUtils;
 
 import org.junit.Assert;
 import org.junit.Test;
