+++ core/server/master/src/test/java/alluxio/master/file/options/CompleteFileOptionsTest.java
@@ -11,7 +11,7 @@
 
 package alluxio.master.file.options;
 
-import alluxio.CommonTestUtils;
+import alluxio.test.utils.CommonTestUtils;
 
 import org.junit.Assert;
 import org.junit.Test;
